# winalog-go/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/system/processes": {
            "get": {
                "description": "返回系统进程列表(Windows)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取进程列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回数量限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ProcessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/system/tasks": {
            "get": {
                "description": "返回系统计划任务列表(Windows)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取计划任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TaskResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/users": {
            "get": {
                "description": "返回系统本地用户列表(Windows)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.ErrorCode"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ProcessInfo": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "string"
                },
                "command_line": {
                    "type": "string"
                },
                "exe": {
                    "type": "string"
                },
                "is_signed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "ppid": {
                    "type": "integer"
                },
                "signature": {
                    "$ref": "#/definitions/api.SignatureInfo"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "api.ProcessResponse": {
            "type": "object",
            "properties": {
                "processes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ProcessInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.SignatureInfo": {
            "type": "object",
            "properties": {
                "issuer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "thumbprint": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_to": {
                    "type": "string"
                }
            }
        },
...
        "api.TaskInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "api.TaskResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.UserInfo": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "home_dir": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UserInfo"
                    }
                }
            }
        },
...
        "types.ErrorCode": {
            "type": "string",
            "enum": [
                "SUCCESS",
                "INTERNAL_ERROR",
                "INVALID_PARAM",
                "INVALID_REQUEST",
                "INVALID_QUERY",
                "NOT_FOUND",
                "EVENT_NOT_FOUND",
                "ALERT_NOT_FOUND",
                "ALERT_ALREADY_RESOLVED",
                "UNAUTHORIZED",
                "PARSE_FAILED",
                "IMPORT_FAILED",
                "FILE_NOT_FOUND",
                "FILE_LOCKED",
                "INVALID_FORMAT",
                "DB_ERROR",
                "DB_READ_ONLY",
                "RULE_INVALID",
                "RULE_NOT_FOUND",
                "RULE_DISABLED",
                "SEARCH_FAILED",
                "RESULT_TOO_LARGE",
                "HASH_MISMATCH",
                "SIGNATURE_INVALID",
                "NOT_SUPPORTED"
            ],
            "x-enum-varnames": [
                "ErrCodeSuccess",
                "ErrCodeInternalError",
                "ErrCodeInvalidParam",
                "ErrCodeInvalidRequest",
                "ErrCodeInvalidQuery",
                "ErrCodeNotFound",
                "ErrCodeEventNotFound",
                "ErrCodeAlertNotFound",
                "ErrCodeAlertAlreadyResolved",
                "ErrCodeUnauthorized",
                "ErrCodeParseFailed",
                "ErrCodeImportFailed",
                "ErrCodeFileNotFound",
                "ErrCodeFileLocked",
                "ErrCodeInvalidFormat",
                "ErrCodeDBError",
                "ErrCodeDBReadOnly",
                "ErrCodeRuleInvalid",
                "ErrCodeRuleNotFound",
                "ErrCodeRuleDisabled",
                "ErrCodeSearchFailed",
                "ErrCodeResultTooLarge",
                "ErrCodeHashMismatch",
                "ErrCodeSignatureInvalid",
                "ErrCodeNotSupported"
            ]
        },
...
    }
}